Rename main to launch
